Second upload
I'm back and my project is being changed. I thought more about layout of my code and I decided to find a way to create more classes and methods. Now the project has 12 classes. Like at the first upload, main class is the starting activity. Next we can find general class of the game called Board.java; In FieldFeatures.java are all settings to making objects from FieldFeatures. The same situation is in the ObjectFeatures.java; Next in GetFeatures.java we can see all objects created in two previous classes. In LevelChooser.java are methods that allow to select which level we have. Now the game select bassic settings for all levels from BasicCreateSettings.java and current settings for the appropriate level from rhigt methods in LevelChooser.java; Now if we have all settings and objects on start the current level will start. This game is the Drag&Drop example so we can find MyTouchListener and MyDragListener classes. When the object will be dropped the Inheritance class will be start. In this class